fix(control-plane): CP-15 use f64 instead of f32 for precision
- Replace all f32 types with f64 in brain engine calculations
- Update signal.rs, config.rs, factors.rs, simulate.rs
- Prevents precision loss in confidence calculations

+++ services/control-plane/src/brain/config.rs
@@ -40,7 +40,7 @@ pub struct IdentityConfig {
     /// Trading mission/style description
     pub mission: String,
     /// Temperament: 0.0 = cautious, 1.0 = bold
-    pub temperament: f32,
+    pub temperament: f64,
 }
 
 impl Default for IdentityConfig {
